Revert "Appi"

+++ Project/src/main/java/mx/LemonTrees/Project/Repository/CargaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import mx.LemonTrees.Project.Model.Carga;
 
 @Repository
-public interface CargaRepository extends CrudRepository<Carga,Long>{
+public interface CargaRepository extends CrudRepository<Carga,Integer>{
 
 }
